refactor LoaderMain to support a constructor with a preconstructed RepositoryManager and a load method using an InputStream
This removes the dependency on Files, so any internal InputStream can be used, including URLs or in-memory streams, for example.

In addition, it can be used with remote repositories if necessary, and new repositories, without modifying LoaderMain using the new constructor.

The old APIs are still intact, including logging as errors, but not throwing RDF parsing and processing exceptions.

The new InputStream load method passes exceptions through, as it is only useful for a single document at a time, so it makes sense to pass exceptions through.

Although the old public APIs are intact, internally there was some rearrangement necessary to support the InputStream load method. I couldn't easily determine a way to keep the basic code for the two public load methods from being duplicated, so there is a small amount of code duplication in this patch.

The getRepositoryManager method was changed to static to support its use inside the constructor hierarchy.